+++ src/helpers/utils.js
@@ -1,0 +1,12 @@
+function isString(object) {
+  return typeof object === "string";
+}
+
+function isArray(object) {
+  return Array.isArray(object);
+}
+
+module.exports = {
+  isString,
+  isArray,
+};