+++ src/rules/element-types.js
@@ -8,6 +8,7 @@ const {
   isMatchElementType,
   elementRulesAllowDependency,
 } = require("../helpers/rules");
+const { ruleElementMessage } = require("../helpers/messages");
 
 function elementRulesAllowDependencyType(element, dependency, options) {
   return elementRulesAllowDependency({
@@ -15,7 +16,31 @@ function elementRulesAllowDependencyType(element, dependency, options) {
     dependency,
     options,
     isMatch: isMatchElementType,
-  }).result;
+  });
+}
+
+function customErrorMessage(message /*, elementCapturedValues, dependencyCapturedValues*/) {
+  // TODO, replace element captured values, including "type", and dependency captured values, including "type"
+  // "Do not import ${dependency.type} from ${file.type} with ${file.elementName}";
+  // TODO, support "message" property in rule settings at first level. Use it if present.
+  return message;
+}
+
+function errorMessage(ruleData, elementCapturedValues) {
+  const ruleReport = ruleData.ruleReport;
+  if (ruleReport.isDefault) {
+    return "Importing elements is disallowed by default. No rule allowing this dependency was found";
+  }
+  if (ruleReport.message) {
+    return customErrorMessage(ruleData, elementCapturedValues);
+  }
+  return `Importing ${ruleElementMessage(
+    ruleReport.disallow,
+    elementCapturedValues
+  )} is not allowed in ${ruleElementMessage(
+    ruleReport.element,
+    elementCapturedValues
+  )}. Disallowed in rule ${ruleReport.index + 1}`;
 }
 
 module.exports = dependencyRule(
@@ -25,18 +50,15 @@ module.exports = dependencyRule(
     schema: rulesOptionsSchema(),
   },
   function ({ dependency, file, node, context, options }) {
-    if (
-      dependency.isLocal &&
-      !dependency.isIgnored &&
-      dependency.type &&
-      !dependency.isInternal &&
-      !elementRulesAllowDependencyType(file, dependency, options)
-    ) {
-      context.report({
-        message: `Usage of '${dependency.type}' is not allowed in '${file.type}'`,
-        node: node,
-        ...dependencyLocation(node, context),
-      });
+    if (dependency.isLocal && !dependency.isIgnored && dependency.type && !dependency.isInternal) {
+      const ruleData = elementRulesAllowDependencyType(file, dependency, options);
+      if (!ruleData.result) {
+        context.report({
+          message: errorMessage(ruleData, file.capturedValues),
+          node: node,
+          ...dependencyLocation(node, context),
+        });
+      }
     }
   }
 );