+++ test/rules/base-pattern/element-types.spec.js
@@ -1,10 +1,8 @@
 const { ELEMENT_TYPES: RULE } = require("../../../src/constants/rules");
 const { SETTINGS, createRuleTester, pathResolvers } = require("../../support/helpers");
+const { ELEMENT_TYPES_DEFAULT_MESSAGE } = require("../../support/messages");
 
 const rule = require(`../../../src/rules/${RULE}`);
-
-const errorMessage = (fileType, dependencyType) =>
-  `Usage of '${dependencyType}' is not allowed in '${fileType}'`;
 
 const settings = SETTINGS.basePattern;
 
@@ -140,7 +138,7 @@ ruleTester.run(RULE, rule, {
       options,
       errors: [
         {
-          message: errorMessage("modules", "modules"),
+          message: ELEMENT_TYPES_DEFAULT_MESSAGE,
           type: "ImportDeclaration",
         },
       ],
@@ -152,7 +150,7 @@ ruleTester.run(RULE, rule, {
       options,
       errors: [
         {
-          message: errorMessage("modules", "modules"),
+          message: ELEMENT_TYPES_DEFAULT_MESSAGE,
           type: "ImportDeclaration",
         },
       ],
@@ -164,7 +162,7 @@ ruleTester.run(RULE, rule, {
       options,
       errors: [
         {
-          message: errorMessage("modules", "components"),
+          message: ELEMENT_TYPES_DEFAULT_MESSAGE,
           type: "ImportDeclaration",
         },
       ],
@@ -176,7 +174,7 @@ ruleTester.run(RULE, rule, {
       options,
       errors: [
         {
-          message: errorMessage("components", "modules"),
+          message: ELEMENT_TYPES_DEFAULT_MESSAGE,
           type: "ImportDeclaration",
         },
       ],
@@ -188,7 +186,7 @@ ruleTester.run(RULE, rule, {
       options,
       errors: [
         {
-          message: errorMessage("components", "modules"),
+          message: ELEMENT_TYPES_DEFAULT_MESSAGE,
           type: "ImportDeclaration",
         },
       ],
@@ -200,7 +198,7 @@ ruleTester.run(RULE, rule, {
       options,
       errors: [
         {
-          message: errorMessage("components", "components"),
+          message: ELEMENT_TYPES_DEFAULT_MESSAGE,
           type: "ImportDeclaration",
         },
       ],
@@ -214,7 +212,7 @@ ruleTester.run(RULE, rule, {
       options,
       errors: [
         {
-          message: errorMessage("components", "components"),
+          message: ELEMENT_TYPES_DEFAULT_MESSAGE,
           type: "ImportDeclaration",
         },
       ],
@@ -226,7 +224,7 @@ ruleTester.run(RULE, rule, {
       options,
       errors: [
         {
-          message: errorMessage("components", "modules"),
+          message: ELEMENT_TYPES_DEFAULT_MESSAGE,
           type: "ImportDeclaration",
         },
       ],
@@ -238,7 +236,7 @@ ruleTester.run(RULE, rule, {
       options,
       errors: [
         {
-          message: errorMessage("components", "modules"),
+          message: ELEMENT_TYPES_DEFAULT_MESSAGE,
           type: "ImportDeclaration",
         },
       ],