+++ test/rules/one-level/invalid-settings.spec.js
@@ -1,14 +1,12 @@
 const { ELEMENT_TYPES: RULE } = require("../../../src/constants/rules");
 const { SETTINGS, createRuleTester, pathResolvers } = require("../../support/helpers");
+const { elementTypesErrorMessage } = require("../../support/messages");
 
 const rule = require(`../../../src/rules/${RULE}`);
 
 const { absoluteFilePath } = pathResolvers("one-level");
 
-const errorMessage = (fileType, dependencyType) =>
-  `Usage of '${dependencyType}' is not allowed in '${fileType}'`;
-
-const test = (settings, options) => {
+const test = (settings, options, errorMessages) => {
   const ruleTester = createRuleTester(settings);
 
   ruleTester.run(RULE, rule, {
@@ -62,7 +60,7 @@ const test = (settings, options) => {
         ],
         errors: [
           {
-            message: errorMessage("helpers", "helpers"),
+            message: elementTypesErrorMessage(errorMessages, 0),
             type: "ImportDeclaration",
           },
         ],
@@ -82,7 +80,8 @@ test(
     {
       default: "disallow",
     },
-  ]
+  ],
+  {}
 );
 
 // no type
@@ -97,7 +96,8 @@ test(
       },
     ],
   },
-  []
+  [],
+  {}
 );
 
 // no valid mode
@@ -111,7 +111,8 @@ test(
       },
     ],
   },
-  []
+  [],
+  {}
 );
 
 // no valid capture
@@ -126,5 +127,6 @@ test(
       },
     ],
   },
-  []
+  [],
+  {}
 );